backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
pytest==8.0.0
pytest-django==4.8.0
gunicorn==21.2.0
whitenoise==6.6.0
PyPDF2==3.0.1
python-docx==1.0.1 